refactor(client): apply organize-components findings to domain edit folder
Practise what the updated skill preaches on this same folder. Pull the pure
server->draft radar-config transforms out of ConfigTab into
config/-radarConfigDrafts.ts so they are unit-testable in isolation (added
tests for padding, sorting, position reindexing, and isAdopted defaults).
Lift the 313-line BlipsTabContainer's state/derivation/handlers into a
blips/-useBlipDrafts hook, leaving the container a slim orchestrator that
wires the hook into the presentational BlipsPanel. Behaviour unchanged.
