+++ packages/client/src/routes/domains.$id.edit.-components/blips/-useBlipDrafts.ts
@@ -1,0 +1,310 @@
+import type { BlipDraft } from "./-BlipsPanel";
+import type { Radar, RadarBlip, TopicForTopicsPage } from "@emstack/types";
+
+import { useEffect, useMemo, useState } from "react";
+
+import { toast } from "sonner";
+
+import { createRadarBlip, deleteRadarBlip, upsertRadarBlip } from "@/utils";
+
+interface UseBlipDraftsArgs {
+  radar: Radar | undefined;
+  topics: TopicForTopicsPage[];
+  domainId: string;
+  onSaved: () => Promise<void>;
+}
+
+let localKeyCounter = 0;
+function nextLocalKey() {
+  localKeyCounter += 1;
+  return `local-${localKeyCounter}`;
+}
+
+// All the state, derivation, and persistence handlers behind the Blips tab.
+// Extracted from the container so -BlipsTab stays a slim orchestrator that just
+// wires this hook into the presentational <BlipsPanel>.
+export function useBlipDrafts({
+  radar,
+  topics,
+  domainId,
+  onSaved,
+}: UseBlipDraftsArgs) {
+  const [newBlipDrafts, setNewBlipDrafts] = useState<BlipDraft[]>([]);
+  const [pendingBlipKey, setPendingBlipKey] = useState<string | null>(null);
+
+  // When the radar changes (e.g., after a save), drop drafts that have been
+  // persisted (their localKey will appear in the new server data via id).
+  useEffect(() => {
+    setNewBlipDrafts(prev =>
+      prev.filter(d => !d.id || !radar?.blips.some(b => b.id === d.id)));
+  }, [radar]);
+
+  const persistedQuadrants = useMemo(
+    () =>
+      (radar?.quadrants ?? []).map(q => ({
+        id: q.id,
+        name: q.name,
+        position: q.position,
+      })),
+    [radar],
+  );
+  const persistedRings = useMemo(
+    () =>
+      (radar?.rings ?? []).map(r => ({
+        id: r.id,
+        name: r.name,
+        position: r.position,
+        isAdopted: r.isAdopted ?? false,
+      })),
+    [radar],
+  );
+  const allConfigPersisted
+    = persistedQuadrants.length > 0 && persistedRings.length > 0;
+
+  const savedBlipsForTable = useMemo(() => radar?.blips ?? [], [radar]);
+
+  const usedTopicIds = useMemo(() => {
+    const set = new Set<string>();
+    savedBlipsForTable.forEach((b) => {
+      if (b.topicId) set.add(b.topicId);
+    });
+    newBlipDrafts.forEach((b) => {
+      if (b.topicId) set.add(b.topicId);
+    });
+    return set;
+  }, [savedBlipsForTable, newBlipDrafts]);
+
+  const topicNameById = useMemo(() => {
+    const map = new Map<string, string>();
+    topics.forEach(t => map.set(t.id, t.name));
+    return map;
+  }, [topics]);
+
+  const topicById = useMemo(() => {
+    const map = new Map<
+      string,
+      { name: string;
+        description?: string | null; }
+    >();
+    topics.forEach(t =>
+      map.set(t.id, {
+        name: t.name,
+        description: t.description,
+      }));
+    return map;
+  }, [topics]);
+
+  function addBlipDraft() {
+    if (persistedQuadrants.length === 0 || persistedRings.length === 0) {
+      toast.error("Add at least one slice and ring first.");
+      return;
+    }
+    setNewBlipDrafts(prev => [
+      ...prev,
+      {
+        topicId: "",
+        description: "",
+        quadrantId: persistedQuadrants[0].id,
+        ringId: persistedRings[0].id,
+        localKey: nextLocalKey(),
+      },
+    ]);
+  }
+
+  function changeBlipTopic(localKey: string, topicId: string) {
+    setNewBlipDrafts(prev =>
+      prev.map(b =>
+        b.localKey === localKey
+          ? {
+            ...b,
+            topicId,
+          }
+          : b));
+  }
+  function changeBlipQuadrant(localKey: string, quadrantId: string) {
+    setNewBlipDrafts(prev =>
+      prev.map(b =>
+        b.localKey === localKey
+          ? {
+            ...b,
+            quadrantId,
+          }
+          : b));
+  }
+  function changeBlipRing(localKey: string, ringId: string) {
+    setNewBlipDrafts(prev =>
+      prev.map(b =>
+        b.localKey === localKey
+          ? {
+            ...b,
+            ringId,
+          }
+          : b));
+  }
+  function changeBlipDescription(localKey: string, description: string) {
+    setNewBlipDrafts(prev =>
+      prev.map(b =>
+        b.localKey === localKey
+          ? {
+            ...b,
+            description,
+          }
+          : b));
+  }
+
+  async function saveBlip(blip: BlipDraft) {
+    if (!blip.topicId) {
+      toast.error("Pick a topic for this blip.");
+      return;
+    }
+    if (!blip.quadrantId || !blip.ringId) {
+      toast.error("Pick a slice and ring.");
+      return;
+    }
+    setPendingBlipKey(blip.localKey);
+    try {
+      const payload = {
+        topicId: blip.topicId,
+        description: blip.description.trim() || null,
+        quadrantId: blip.quadrantId,
+        ringId: blip.ringId,
+      };
+      if (blip.id) {
+        await upsertRadarBlip(domainId, blip.id, payload);
+      }
+      else {
+        await createRadarBlip(domainId, payload);
+      }
+      setNewBlipDrafts(prev =>
+        prev.filter(b => b.localKey !== blip.localKey));
+      await onSaved();
+      toast.success("Blip saved.");
+    }
+    catch {
+      toast.error("Failed to save blip.");
+    }
+    finally {
+      setPendingBlipKey(null);
+    }
+  }
+
+  async function removeBlip(blip: BlipDraft) {
+    if (blip.id) {
+      setPendingBlipKey(blip.localKey);
+      try {
+        await deleteRadarBlip(domainId, blip.id);
+        await onSaved();
+      }
+      catch {
+        toast.error("Failed to delete blip.");
+        setPendingBlipKey(null);
+        return;
+      }
+      setPendingBlipKey(null);
+    }
+    setNewBlipDrafts(prev =>
+      prev.filter(b => b.localKey !== blip.localKey));
+  }
+
+  async function handleTableSave(
+    blip: RadarBlip,
+    patch: {
+      quadrantId: string | null;
+      ringId: string | null;
+      description: string | null;
+      isIgnored: boolean;
+    },
+  ) {
+    try {
+      await upsertRadarBlip(domainId, blip.id, {
+        topicId: blip.topicId,
+        quadrantId: patch.quadrantId,
+        ringId: patch.ringId,
+        description: patch.description,
+        isIgnored: patch.isIgnored,
+      });
+      await onSaved();
+      toast.success("Blip saved.");
+    }
+    catch {
+      toast.error("Failed to save blip.");
+      throw new Error("save failed");
+    }
+  }
+
+  async function handleTableRemove(blip: RadarBlip) {
+    try {
+      await deleteRadarBlip(domainId, blip.id);
+      await onSaved();
+    }
+    catch {
+      toast.error("Failed to delete blip.");
+      throw new Error("delete failed");
+    }
+  }
+
+  async function handleTableBulkSave(
+    ids: string[],
+    patch: { quadrantId?: string;
+      ringId?: string; },
+  ) {
+    if (ids.length === 0) return;
+    if (patch.quadrantId === undefined && patch.ringId === undefined) return;
+    const blipById = new Map<string, RadarBlip>();
+    (radar?.blips ?? []).forEach(b => blipById.set(b.id, b));
+    let updated = 0;
+    let failed = 0;
+    for (const id of ids) {
+      const blip = blipById.get(id);
+      if (!blip) {
+        failed += 1;
+        continue;
+      }
+      try {
+        await upsertRadarBlip(domainId, id, {
+          topicId: blip.topicId,
+          quadrantId: patch.quadrantId ?? blip.quadrantId,
+          ringId: patch.ringId ?? blip.ringId,
+          description: blip.description ?? null,
+          isIgnored: blip.isIgnored ?? false,
+        });
+        updated += 1;
+      }
+      catch {
+        failed += 1;
+      }
+    }
+    await onSaved();
+    if (updated > 0 && failed === 0) {
+      toast.success(`Updated ${updated} blip${updated === 1 ? "" : "s"}.`);
+    }
+    else if (updated > 0 && failed > 0) {
+      toast.error(`Updated ${updated}, failed ${failed}.`);
+    }
+    else {
+      toast.error("Failed to update blips.");
+    }
+  }
+
+  return {
+    allConfigPersisted,
+    savedBlipsForTable,
+    newBlipDrafts,
+    persistedQuadrants,
+    persistedRings,
+    usedTopicIds,
+    pendingBlipKey,
+    topicById,
+    topicNameById,
+    addBlipDraft,
+    changeBlipTopic,
+    changeBlipQuadrant,
+    changeBlipRing,
+    changeBlipDescription,
+    saveBlip,
+    removeBlip,
+    handleTableSave,
+    handleTableRemove,
+    handleTableBulkSave,
+  };
+}
